[Cleanup] Part 1
This is the first of several cleanup commits, specifically directed towards the dependencies.

This commits removes the old, deprecate precompiled openssl and mysql libraries, as well as started restructuring of shared and deps folder.

This has not been tested on Linux yet, and the pre-built SLN files have not yet been updated.