Stranglethorn Vale sleeping creatures script
SD2 script added, that will make certain creatures in Stranglethorn Vale
sleep during the hours of 9pm and 5am.

Only the panthers and tigers were found to have a sleeping
animation/stance, so only they are affected.

Also, World.h has been included in precompiled.h, in order to have
access to the time of day.