Breath timer
Base duration of breath should be 60 seconds in Vanilla and TBC, not 3
minutes as it was changed to in patch 3.0.2 for Wrath (and how it
currently is on cmangos classic)

http://wowwiki.wikia.com/wiki/Breath?oldid=335923

http://wowwiki.wikia.com/wiki/Breath

It even states in the uncached, current version of the wiki, my last
link, which patch it was changed in, and what the duration was prior to
that; 60 seconds.

TODO: The forsaken racial trait should be changed to reflect this,
meaning that undead should have 3 minutes base, regardless of what they
had before, as my first link clearly states.