Take takes the implementation of Mangos core One Unit::IsNonMeleeSpellCasted() and use it in Mangos core Zero.
I don't think this is a proper fix, but it does allow a frost bolt to be cast while the previous one is in flight to the target, and the mobs still cast their spells as expected.

With this cahnge it appears the warlock imp now casts fireballs, where it was only doing melee before, but it is casting them way too fast so I don't think this is a complete fix.

This in regards to this bug in the tracker: https://www.getmangos.eu/bug-tracker/mangos-zero/unable-to-cast-next-frost-bolt-while-current-one-is-traveling-r1726/