[Part 1] Updating all workflow files - Updated workflow for code styling - Updated linux build file
For the code styling, I have added 6 new checks that will check code for a few things. Refer to the PR description area for more details.

Note: Making sure Linux is using the same version of OpenSSL as the Windows workflow build. So we are consistent with that package if problems happen. On the workflow side, default version is 3.0.2 from what I been seeing thus far.